requirements.txt
urllib3==2.2.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
googletrans==3.1.0a0 